home_monitor/requirements.txt
numpy==1.23.4
pandas==1.5.0
plotly
speedtest-cli
